% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{barchart}
\alias{barchart}
\title{Bar Chart}
\usage{
barchart(data)
}
\arguments{
\item{data}{dataframe.}
}
\description{
This function allows you to interactively create a scatterplot and acess the ggplot syntax.
}
\examples{
data(cars)
barchart(cars)
}
\keyword{barchart}
