% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Histogram}
\usage{
histogram(data)
}
\arguments{
\item{data}{dataframe.}
}
\description{
This function allows you to interactively create a scatterplot and acess the ggplot syntax.
}
\examples{
data(cars)
histogram(cars)
}
\keyword{histogram}
