% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot}
\usage{
scatterplot(data)
}
\arguments{
\item{data}{dataframe.}
}
\description{
This function allows you to interactively create a scatterplot and acess the ggplot syntax.
}
\examples{
data(cars)
scatterplot(cars)
}
\keyword{scatterplot}
